Fix an RTC issue with losing contents at power off
When powering the circuit from VCC normally, the contents of the RTC
would fail to be held if VCC was the source of power during normal
operation.  the timekeeping was always powered by Vbat, there wasn't
an issue.  The problem is that the voltage into VCC cannot drop
faster than .5V/ms or the RTC can't switch in time.

Add a diode and capacitor to VCC to slow the power down rate.

Signed-off-by: Corey Minyard <corey@minyard.net>