Add a buffer from OTHER_HW_POWER_OFF_N to the CPU
That particular line is also hooked to the power disable on the other
board.  If you reset the other board, when the other board is powered
off, the CPU appears to pull the OTHER_HW_POWER_OFF_N down pretty
strongly causing a reset on the first board.

Add a buffer going into the CPU to fix the issue.

Signed-off-by: Corey Minyard <corey@minyard.net>