Fix a reset issue with ACTIVE_N
Like the OTHER_HW_POWER_OFF_N in a previous commit, the ACTIVE_N line
was pulled low when the CPU was in reset, which could mess up fault
tolerance because ACTIVE1 controls the RF output switch.  Add a
transistor there and switch it to positive logic, ACTIVE instead.  FAULT
doesn't seem to have the issue and the rest of the lines shouldn't
matter as they are only outputs from the other board.

Signed-off-by: Corey Minyard <corey@minyard.net>